src/requirements.txt
python-dotenv
discord
spotipy